Remove ignored file. Add check before rm in watch script

+++ css/watch_css.sh
@@ -19,4 +19,6 @@ sass -t compressed --watch "$sassfile":"$intermed" &\
 trap - SIGINT
 
 # Done with intermediate file
-rm "$intermed"
+if [ -f "$intermed" ]; then
+    rm "$intermed"
+fi
